requirements.txt
aiogram==3.13.1
aiohttp>=3.9.5
openpyxl==3.1.2
aiofiles>=23.2.1
aiosmtplib==2.0.0
psutil>=5.9.0  # or whichever version you need
python-dotenv==1.0.0
redis==4.5.5
celery==5.2.7
python-multipart==0.0.6
loguru==0.7.0
prometheus_client
prometheus_client  #e1

